add: add margin and padding to `NewNoteworthy` and create item for each #12
