+++ Podcast-CloneCoding/Podcast-CloneCoding/View/FeaturedChannels/FeaturedChannelsRow.swift
@@ -18,16 +18,14 @@ struct FeaturedChannelsRow: View {
         VStack(alignment: .leading) {
             Divider()
                 .padding(EdgeInsets(top: 0, leading: 20, bottom: 0, trailing: 20))
-            VStack {
-                Text("Featured Channels")
-                    .font(.title2)
-                    .bold()
-            }
-            .padding(EdgeInsets(top: 0, leading: 16, bottom: 0, trailing: 0))
+            Text("Featured Channels")
+                .font(.title2)
+                .bold()
+                .padding(EdgeInsets(top: 0, leading: 20, bottom: 0, trailing: 0))
             ScrollView(.horizontal, showsIndicators: false) {
                 VStack {
                     HStack(spacing: 12) {
-                        Spacer(minLength: 0)
+                        Spacer()
                         ForEach(featuredChannels) { featuredChannel in
                             FeaturedChannelsItem(featuredChannel: featuredChannel)
                                 .frame(width: 250, height: 350, alignment: .leading)
@@ -35,7 +33,7 @@ struct FeaturedChannelsRow: View {
                                 .cornerRadius(10)
                                 .shadow(color: .black.opacity(0.2), radius: 3, x: 0, y: 2)
                         }
-                        Spacer(minLength: 0)
+                        Spacer()
                     }
                     Spacer()
                 }
