+++ Podcast-CloneCoding/Podcast-CloneCoding/View/Tab/Browse.swift
@@ -27,7 +27,9 @@ struct Browse: View {
                     // FeaturedRow's item height is 310.
                     FeaturedRow(featuredes: modelData.featured)
                         .frame(height: 330)
+                    // NewNotewrothyRow's item height is 210.
                     NewNoteworthyRow(newNoteworthys: modelData.newNoteworthy)
+                        .frame(height: 280)
                     // FeaturedChannelsRow's item height is 350.
                     FeaturedChannelsRow(featuredChannels: modelData.featuredChannels)
                         .frame(height: 420)
